requirements.txt
scikit-learn==1.2.0
matplotlib==3.6.2
black==22.12.0
black[jupyter]==22.12.0
isort==5.11.4
plotly==5.11.0
nbformat==5.7.1
pandas==1.5.2
seaborn==0.11.2
mldesigner==0.1.0b7
azure-ai-ml==1.3.0
azureml-core==1.52.0
azure-identity==1.13.0